+++ docs/src/app_types/tutorial005.py
@@ -1,0 +1,6 @@
+from corva import Api, Cache, ScheduledDepthEvent, scheduled  # <1>
+
+
+@scheduled  # <3>
+def scheduled_app(event: ScheduledDepthEvent, api: Api, cache: Cache):  # <2>
+    return 'Hello, World!'  # <4>