+++ docs/src/testing/tutorial003.py
@@ -1,13 +1,15 @@
-from corva import Api, Cache, ScheduledEvent, scheduled
+from corva import Api, Cache, ScheduledDataTimeEvent, scheduled
 
 
 @scheduled
-def scheduled_app(event: ScheduledEvent, api: Api, cache: Cache):  # <.>
+def scheduled_app(event: ScheduledDataTimeEvent, api: Api, cache: Cache):  # <.>
     return 'Hello, World!'
 
 
 def test_scheduled_app(app_runner):  # <.>
-    event = ScheduledEvent(asset_id=0, start_time=0, end_time=0, company_id=0)  # <.>
+    event = ScheduledDataTimeEvent(
+        asset_id=0, start_time=0, end_time=0, company_id=0
+    )  # <.>
 
     result = app_runner(scheduled_app, event=event)  # <.>