+++ src/corva/__init__.py
@@ -1,7 +1,11 @@
 from .api import Api
 from .handlers import scheduled, stream, task
 from .logger import CORVA_LOGGER as Logger
-from .models.scheduled import ScheduledEvent
+from .models.scheduled.scheduled import (
+    ScheduledDataTimeEvent,
+    ScheduledDepthEvent,
+    ScheduledNaturalTimeEvent,
+)
 from .models.stream.stream import (
     StreamDepthEvent,
     StreamDepthRecord,
@@ -11,17 +15,19 @@ from .models.stream.stream import (
 from .models.task import TaskEvent
 from .state.redis_state import RedisState as Cache
 
-__all__ = [
-    'Api',
-    'Cache',
-    'Logger',
-    'ScheduledEvent',
-    'StreamDepthEvent',
-    'StreamDepthRecord',
-    'StreamTimeEvent',
-    'StreamTimeRecord',
-    'TaskEvent',
-    'scheduled',
-    'stream',
-    'task',
-]
+
+def __getattr__(name):
+    import warnings
+
+    if name == "ScheduledEvent":
+        warnings.warn(
+            "The corva.ScheduledEvent class is deprecated "
+            "and will be removed from corva in the next major version. "
+            "Import corva.ScheduledDataTimeEvent instead.",
+            FutureWarning,
+            stacklevel=2,
+        )
+
+        return ScheduledDataTimeEvent
+
+    raise AttributeError(f"module {__name__!r} has no attribute {name!r}")