+++ tests/test_docs/test_app_types.py
@@ -1,12 +1,21 @@
 from corva import (
-    ScheduledEvent,
+    ScheduledDataTimeEvent,
+    ScheduledDepthEvent,
+    ScheduledNaturalTimeEvent,
     StreamDepthEvent,
     StreamDepthRecord,
     StreamTimeEvent,
     StreamTimeRecord,
     TaskEvent,
 )
-from docs.src.app_types import tutorial001, tutorial002, tutorial003, tutorial004
+from docs.src.app_types import (
+    tutorial001,
+    tutorial002,
+    tutorial003,
+    tutorial004,
+    tutorial005,
+    tutorial006,
+)
 
 
 def test_tutorial001(app_runner):
@@ -26,7 +35,7 @@ def test_tutorial002(app_runner):
 
 
 def test_tutorial003(app_runner):
-    event = ScheduledEvent(asset_id=0, start_time=0, end_time=0, company_id=0)
+    event = ScheduledDataTimeEvent(asset_id=0, start_time=0, end_time=0, company_id=0)
 
     assert app_runner(tutorial003.scheduled_app, event) == 'Hello, World!'
 
@@ -35,3 +44,24 @@ def test_tutorial004(app_runner):
     event = TaskEvent(asset_id=0, company_id=0)
 
     assert app_runner(tutorial004.task_app, event) == 'Hello, World!'
+
+
+def test_tutorial005(app_runner):
+    event = ScheduledDepthEvent(
+        asset_id=0,
+        company_id=0,
+        top_depth=0.0,
+        bottom_depth=1.0,
+        log_identifier='',
+        interval=1.0,
+    )
+
+    assert app_runner(tutorial005.scheduled_app, event) == 'Hello, World!'
+
+
+def test_tutorial006(app_runner):
+    event = ScheduledNaturalTimeEvent(
+        asset_id=0, company_id=0, schedule_start=0, interval=1
+    )
+
+    assert app_runner(tutorial006.scheduled_app, event) == 'Hello, World!'