+++ CHANGELOG.md
@@ -6,8 +6,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Subtypes of scheduled event:
+`ScheduledDataTimeEvent`, `ScheduledDepthEvent` and `ScheduledNaturalTimeEvent`  
 
-## [1.1.0]
+### Deprecated
+- `ScheduledEvent` usage
+
+
+## [1.1.0] - 2021-07-19
 
 ### Added
 - Ability to pass custom logging handlers